+++ src/components/user-progrees/index.tsx
@@ -1,0 +1,3 @@
+import { ProgressPage } from "./user-progress-page";
+
+export { ProgressPage };
