+++ src/components/user-progrees/user-progress-page.tsx
@@ -1,0 +1,57 @@
+import { descriptionStyle, pageTitleStyle } from "GlobalStyles";
+import React from "react";
+import { StockOutlined } from "@ant-design/icons";
+import { UserProgress } from "./user-progress";
+
+export { ProgressPage };
+
+type FlexCenterDivPropsWithTitle = {
+  title: string | JSX.Element;
+  children: JSX.Element;
+};
+const PageWithTitle: React.FC<FlexCenterDivPropsWithTitle> = ({
+  title,
+  children,
+}) => {
+  return (
+    <div
+      style={{
+        textAlign: "center",
+        marginTop: "3rem",
+        marginBottom: "15rem",
+      }}
+    >
+      <section
+        style={{
+          ...pageTitleStyle,
+          marginBottom: "1rem",
+        }}
+      >
+        {title}
+      </section>
+      <section
+        style={{
+          ...descriptionStyle,
+          padding: "0 5rem",
+          textAlign: "center",
+        }}
+      >
+        {children}
+      </section>
+    </div>
+  );
+};
+
+const ProgressPage: React.FC = () => {
+  return (
+    <PageWithTitle
+      title={
+        <>
+          Your progress <StockOutlined />
+        </>
+      }
+    >
+      <UserProgress />
+    </PageWithTitle>
+  );
+};
